lambda/requirements.txt
firebolt-sdk>=1.0.0
